3.1.6 Przeniesienie klasy IntArrayStack
